Adição dos controllers, para abstrair as rotas
