+++ src/modules/users/infra/http/controllers/UserAvatarController.ts
@@ -1,0 +1,19 @@
+import { Request, Response } from 'express';
+
+import { container } from 'tsyringe';
+
+import UpdateUserAvatarService from '@modules/users/services/UpdateUserAvatarService';
+
+export default class UserAvatarController {
+  public async update(request: Request, response: Response): Promise<Response> {
+    const updateUserAvatarService = container.resolve(UpdateUserAvatarService);
+    const user = await updateUserAvatarService.execute({
+      user_id: request.user.id,
+      avatarFilename: request.file.filename,
+    });
+
+    delete user.password;
+
+    return response.json(user);
+  }
+}
